Fixed issue #2 Added feature for cleanup email to choose which emails to delete. Also updated tests

+++ django_mail_admin/management/commands/cleanup_email.py
@@ -3,7 +3,7 @@ import datetime
 from django.core.management.base import BaseCommand
 from django.utils.timezone import now
 
-from django_mail_admin.models import OutgoingEmail
+from django_mail_admin.models import OutgoingEmail, IncomingEmail
 
 
 class Command(BaseCommand):
@@ -15,11 +15,25 @@ class Command(BaseCommand):
                             default=90,
                             help="Cleanup mails older than this many days, defaults to 90."
                             )
+        parser.add_argument('-i', '--incoming',
+                            type=bool,
+                            default=False,
+                            help="Cleanup incoming mails (with attachments), defaults to False")
+        parser.add_argument('-o', '--outgoing',
+                            type=bool,
+                            default=True,
+                            help="Cleanup outgoing mails, defaults to False")
 
-    def handle(self, verbosity, days, **options):
+    def handle(self, verbosity, days, incoming, outgoing, **options):
         # Delete mails and their related logs and queued created before X days
-
+        if not incoming and not outgoing:
+            print("Please select either incoming or outgoing emails to delete. Exiting...")
         cutoff_date = now() - datetime.timedelta(days)
-        count = OutgoingEmail.objects.filter(created__lt=cutoff_date).count()
-        OutgoingEmail.objects.only('id').filter(created__lt=cutoff_date).delete()
-        print("Deleted {0} mails created before {1} ".format(count, cutoff_date))
+        if outgoing:
+            count_outgoing = OutgoingEmail.objects.filter(created__lt=cutoff_date).count()
+            OutgoingEmail.objects.only('id').filter(created__lt=cutoff_date).delete()
+            print("Deleted {0} outgoing mails created before {1} ".format(count_outgoing, cutoff_date))
+        if incoming:
+            count_incoming = IncomingEmail.objects.filter(processed__lt=cutoff_date).count()
+            IncomingEmail.objects.only('id').filter(processed__lt=cutoff_date).delete()
+            print("Deleted {0} incoming mails processed before {1} ".format(count_incoming, cutoff_date))
